update ui code for fixing bugs and upgrade Clarity version to 0.8.7

+++ src/ui_ng/src/app/repository/repository.component.ts
@@ -5,11 +5,11 @@ import { RepositoryService } from './repository.service';
 import { Repository } from './repository';
 
 import { MessageService } from '../global-message/message.service';
-import { AlertType, DeletionTargets } from '../shared/shared.const';
+import { AlertType, ConfirmationState, ConfirmationTargets } from '../shared/shared.const';
 
 
-import { DeletionDialogService } from '../shared/deletion-dialog/deletion-dialog.service';
-import { DeletionMessage } from '../shared/deletion-dialog/deletion-message';
+import { ConfirmationDialogService } from '../shared/confirmation-dialog/confirmation-dialog.service';
+import { ConfirmationMessage } from '../shared/confirmation-dialog/confirmation-message';
 import { Subscription } from 'rxjs/Subscription';
 
 import { State } from 'clarity-angular';
@@ -20,8 +20,10 @@ const repositoryTypes = [
 ];
 
 @Component({
+  moduleId: module.id,
   selector: 'repository',
-  templateUrl: 'repository.component.html'
+  templateUrl: 'repository.component.html',
+  styleUrls: ['./repository.component.css']
 })
 export class RepositoryComponent implements OnInit {
   changedRepositories: Repository[];
@@ -43,24 +45,28 @@ export class RepositoryComponent implements OnInit {
     private route: ActivatedRoute,
     private repositoryService: RepositoryService,
     private messageService: MessageService,
-    private deletionDialogService: DeletionDialogService
+    private deletionDialogService: ConfirmationDialogService
   ) {
     this.subscription = this.deletionDialogService
-        .deletionConfirm$
-        .subscribe(
-          message=>{
-            let repoName = message.data;
-            this.repositoryService
-                .deleteRepository(repoName)
-                .subscribe(
-                  response=>{
-                    this.refresh();
-                    console.log('Successful deleted repo:' + repoName);
-                  },
-                  error=>this.messageService.announceMessage(error.status, 'Failed to delete repo:' + repoName, AlertType.DANGER)
-                );
-          }
-        );
+      .confirmationConfirm$
+      .subscribe(
+      message => {
+        if (message &&
+          message.source === ConfirmationTargets.REPOSITORY &&
+          message.state === ConfirmationState.CONFIRMED) {
+          let repoName = message.data;
+          this.repositoryService
+            .deleteRepository(repoName)
+            .subscribe(
+            response => {
+              this.refresh();
+              console.log('Successful deleted repo:' + repoName);
+            },
+            error => this.messageService.announceMessage(error.status, 'Failed to delete repo:' + repoName, AlertType.DANGER)
+            );
+        }
+      }
+      );
   }
 
   ngOnInit(): void {
@@ -71,43 +77,45 @@ export class RepositoryComponent implements OnInit {
   }
 
   ngOnDestroy(): void {
-    if(this.subscription) {
+    if (this.subscription) {
       this.subscription.unsubscribe();
     }
   }
 
   retrieve(state?: State) {
-    if(state) {
+    if (state) {
       this.page = state.page.to + 1;
     }
     this.repositoryService
-        .listRepositories(this.projectId, this.lastFilteredRepoName, this.page, this.pageSize)
-        .subscribe(
-          response=>{
-            this.totalRecordCount = response.headers.get('x-total-count');
-            this.totalPage = Math.ceil(this.totalRecordCount / this.pageSize);
-            console.log('TotalRecordCount:' + this.totalRecordCount + ', totalPage:' + this.totalPage);
-            this.changedRepositories=response.json();
-          },
-          error=>this.messageService.announceMessage(error.status, 'Failed to list repositories.', AlertType.DANGER)
-        );
+      .listRepositories(this.projectId, this.lastFilteredRepoName, this.page, this.pageSize)
+      .subscribe(
+      response => {
+        this.totalRecordCount = response.headers.get('x-total-count');
+        this.totalPage = Math.ceil(this.totalRecordCount / this.pageSize);
+        console.log('TotalRecordCount:' + this.totalRecordCount + ', totalPage:' + this.totalPage);
+        this.changedRepositories = response.json();
+      },
+      error => this.messageService.announceMessage(error.status, 'Failed to list repositories.', AlertType.DANGER)
+      );
   }
 
   doFilterRepositoryByType(type: string) {
-    this.currentRepositoryType = this.repositoryTypes.find(r=>r.key == type);
+    this.currentRepositoryType = this.repositoryTypes.find(r => r.key == type);
   }
-  
+
   doSearchRepoNames(repoName: string) {
     this.lastFilteredRepoName = repoName;
     this.retrieve();
-   
+
   }
 
   deleteRepo(repoName: string) {
-    let message = new DeletionMessage(
-      'REPOSITORY.DELETION_TITLE_REPO', 
-      'REPOSITORY.DELETION_SUMMARY_REPO', 
-      repoName, repoName, DeletionTargets.REPOSITORY);
+    let message = new ConfirmationMessage(
+      'REPOSITORY.DELETION_TITLE_REPO',
+      'REPOSITORY.DELETION_SUMMARY_REPO',
+      repoName,
+      repoName,
+      ConfirmationTargets.REPOSITORY);
     this.deletionDialogService.openComfirmDialog(message);
   }
